refactor(src): merge 5 groups of duplicated helpers into shared homes
Five sets of functions had byte-identical or near-identical bodies that
differed only by one or two parameters. Each is now one shared function:

- `conc_sweep._build_comparison` / `breakdown.collectors.kernels.
  ._conc_sweep_pair_comparison` were the same 57-line CONC outer-join
  (the collector copy existed only to stay import-light) ->
  `common.gain_math.conc_pair_comparison`, which is already stdlib-only.
- `credentials._{validate,resolve}_{critic,robustness}_agent_*` differed
  only by the agent package name driving the module path, the
  `<AGENT>_AGENT_*` env vars, and the `--<agent>-mock` hint ->
  `_resolve_agent_root(agent)` / `_validate_agent_runtime(root, agent=)`.
- `sweep._pareto_front` / `_geak_sweep._pareto_front` differed only by
  which latency metric they minimize (e2el vs ttft, because GEAK's
  bench_summary.json has no e2el) -> `_grid_base.pareto_front(entries,
  latency_key=)`.
- `kernel_node_ops._atomic_write_bytes` / `kernel_patch_multinode.
  ._atomic_write_bytes` were identical -> `atomic_write_bytes` in
  `patch_path_safety.py`, which both pod scripts already import and
  which both shipping paths already bundle alongside them.
- `gbrain_page_client._iter_sse_objects` / `gbrain_remote_client.
  ._iter_sse_objects` were the same MCP/SSE decoder ->
  `common.jsonio.iter_sse_objects`.

Duplicate names are deleted rather than wrapped; call sites, `__all__`,
and the duplicated tests are updated or collapsed (the two pareto tests
became one parametrized over `latency_key`). Net production reduction:
-178 lines (+267/-445, measured via `git diff --numstat` over non-test
files); the saving is from collapsing duplicated code paths, not from
trimming comments or docstrings. Behavior unchanged: the full suite has
the same 14 pre-existing failures before and after.
