+++ src/hyperloom/inference_optimizer/cli/__init__.py
@@ -63,11 +63,9 @@ from .credentials import (
     _CLAUDE_ALLOWED_MODELS as _CLAUDE_ALLOWED_MODELS,
     _CATALOG_RETRY_DELAYS_SEC as _CATALOG_RETRY_DELAYS_SEC,
     _CRITIC_AGENT_ROOT_ENV as _CRITIC_AGENT_ROOT_ENV,
-    _resolve_critic_agent_root as _resolve_critic_agent_root,
-    _validate_critic_agent_runtime as _validate_critic_agent_runtime,
     _ROBUSTNESS_AGENT_ROOT_ENV as _ROBUSTNESS_AGENT_ROOT_ENV,
-    _resolve_robustness_agent_root as _resolve_robustness_agent_root,
-    _validate_robustness_agent_runtime as _validate_robustness_agent_runtime,
+    _resolve_agent_root as _resolve_agent_root,
+    _validate_agent_runtime as _validate_agent_runtime,
 )
 from .multi_node import (
     _provision_multi_node_rayjob_stack as _provision_multi_node_rayjob_stack,
@@ -1878,7 +1876,7 @@ async def _run_optimize(args: argparse.Namespace) -> int:
         critic_choice,
         codex_follows_claude=codex_follows_claude,
     ):
-        critic_agent_root = _resolve_critic_agent_root()
+        critic_agent_root = _resolve_agent_root("critic")
         if critic_agent_root is None:
             print(
                 f"ERROR: --critic-agent selected but critic-agent runtime not "
@@ -1890,7 +1888,7 @@ async def _run_optimize(args: argparse.Namespace) -> int:
                 file=sys.stderr,
             )
             sys.exit(2)
-        _validate_critic_agent_runtime(critic_agent_root)
+        _validate_agent_runtime(critic_agent_root, agent="critic")
         if critic_kb_mode == "live" and not os.environ.get("KB_BASE_URL"):
             print(
                 "ERROR: CRITIC_KB_CLIENT_MODE=live but KB_BASE_URL is not "
@@ -1907,7 +1905,7 @@ async def _run_optimize(args: argparse.Namespace) -> int:
     robustness_agent_root: Path | None = None
     robustness_options = _build_robustness_options(args)
     if robustness_choice == "agent":
-        robustness_agent_root = _resolve_robustness_agent_root()
+        robustness_agent_root = _resolve_agent_root("robustness")
         if robustness_agent_root is None:
             print(
                 f"ERROR: --robustness-agent selected but robustness-agent "
@@ -1919,7 +1917,7 @@ async def _run_optimize(args: argparse.Namespace) -> int:
                 file=sys.stderr,
             )
             sys.exit(2)
-        _validate_robustness_agent_runtime(robustness_agent_root)
+        _validate_agent_runtime(robustness_agent_root, agent="robustness")
 
     backends = _build_backends(
         claude_model=args.claude_model,
